quito rescue de exception [skip actions]

+++ app/uploaders/file_uploader.rb
@@ -39,12 +39,10 @@ class FileUploader < CarrierWave::Uploader::Base
     end
 
     def delete_empty_upstream_dirs
-      Dir.delete(store_dir)
+      Dir.delete(store_dir) if Dir.empty?(store_dir)
 
       parent_dir = File.dirname(store_dir)
 
-      Dir.delete(parent_dir)
-    rescue SystemCallError
-      true # nothing, the dir is not empty
+      Dir.delete(parent_dir) if Dir.empty?(parent_dir)
     end
 end
