+++ app/uploaders/image_uploader.rb
@@ -62,12 +62,10 @@ class ImageUploader < CarrierWave::Uploader::Base
     end
 
     def delete_empty_upstream_dirs
-      Dir.delete(store_dir)
+      Dir.delete(store_dir) if Dir.exist?(store_dir) && Dir.empty?(store_dir)
 
       parent_dir = File.dirname(store_dir)
 
-      Dir.delete(parent_dir)
-    rescue SystemCallError
-      true # nothing, the dir is not empty
+      Dir.delete(parent_dir) if Dir.exist?(parent_dir) && Dir.empty?(parent_dir)
     end
 end
